test: useMakeOrEnterRoom hook 테스트 추가 #152
