+++ frontend/src/constants/url.ts
@@ -27,8 +27,6 @@ export const MOCK_API_URL = {
   moveNextRound: `${BASE_URL}/api/balances/rooms/:roomId/next-round`,
   finalResult: `${BASE_URL}/api/balances/rooms/:roomId/final`,
   room: `${BASE_URL}/api/balances/rooms`,
-  enterRoom: `${BASE_URL}/api/balances/rooms/:roomId/members`,
-  getRoomInfo: `${BASE_URL}/api/balances/rooms/:roomId`,
-  startGame: `${BASE_URL}/api/balances/rooms/:roomId/start`,
+  roomMembers: `${BASE_URL}/api/balances/rooms/:roomId`,
   roundVoteIsFinished: `${BASE_URL}/api/balances/rooms/:roomId/contents/:contentId/vote-finished`,
 };
